query_plot

+++ lib/bencher_rank/src/db.rs
@@ -1,9 +1,9 @@
 use crate::Rank;
 
-impl<DB> diesel::serialize::ToSql<diesel::sql_types::Integer, DB> for Rank
+impl<DB> diesel::serialize::ToSql<diesel::sql_types::BigInt, DB> for Rank
 where
     DB: diesel::backend::Backend,
-    i64: diesel::serialize::ToSql<diesel::sql_types::Integer, DB>,
+    i64: diesel::serialize::ToSql<diesel::sql_types::BigInt, DB>,
 {
     fn to_sql<'b>(
         &'b self,
@@ -13,10 +13,10 @@ where
     }
 }
 
-impl<DB> diesel::deserialize::FromSql<diesel::sql_types::Integer, DB> for Rank
+impl<DB> diesel::deserialize::FromSql<diesel::sql_types::BigInt, DB> for Rank
 where
     DB: diesel::backend::Backend,
-    i64: diesel::deserialize::FromSql<diesel::sql_types::Integer, DB>,
+    i64: diesel::deserialize::FromSql<diesel::sql_types::BigInt, DB>,
 {
     fn from_sql(bytes: DB::RawValue<'_>) -> diesel::deserialize::Result<Self> {
         i64::from_sql(bytes).map(Self)
